Move RangeEnumerationExtensions to Extensions directory
