--- Assets/Scripts/UnityUtils.Examples/Enumeration.meta
@@ -1,3 +1,0 @@
-fileFormatVersion: 2
-guid: 18ab976e3655409cbefad0f3c63c8db7
-timeCreated: 1669536561
