--- Assets/Scripts/UnityUtils.Tests/Enumeration.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 5518e617068954fe0b1aaba5526bf0f8
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
