--- Assets/Scripts/UnityUtils/Enumeration.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: d208b5fb8d6fd4f9d83b996b26daafd1
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
